refactor : 패키지 구조 및 네이밍 수정
