+++ src/main/java/be/ddd/infra/filter/Auth0JwtVerifier.java
@@ -1,4 +1,4 @@
-package be.ddd.api.login;
+package be.ddd.infra.filter;
 
 import com.auth0.jwk.JwkProvider;
 import com.auth0.jwk.JwkProviderBuilder;
